Copy / paste actions with Ctrl-Shift-C/V shortcuts (as for gnome-terminal).
They don't integrate with the Eclipse actions because of the key
binding difficulties that result.

For now copy has no enablement as I can't find a suitable way
to do this from JTerminalPane. Terminator itself has no enablement
for either action.

Added a view menu with the actions but really I want a context
menu - will document that struggle in TODO.